+++ packages/plugin-gallery/web/src/components/gallery-controls/gallery-image-settings.js
@@ -155,6 +155,7 @@ class ImageSettings extends Component {
                     className={styles.galleryImageSettings_image}
                     src={this.getMediaUrl(image)}
                     theme={theme}
+                    errorMsg={image.errorMsg}
                   />
                   <div
                     className={classNames(styles.galleryImageSettings_nav, {
@@ -279,6 +280,7 @@ ImageSettings.propTypes = {
   image: PropTypes.shape({
     url: PropTypes.string.isRequired,
     metadata: PropTypes.object.isRequired,
+    errorMsg: PropTypes.string,
   }).isRequired,
   onCancel: PropTypes.func.isRequired,
   onSave: PropTypes.func.isRequired,