+++ CHANGELOG.md
@@ -17,6 +17,8 @@
   </summary>
 
 ### :rocket: New Feature
+- `gallery`
+  - [#1373](https://github.com/wix-incubator/rich-content/pull/1373) gallery plugin upload error handling
 - `file-upload`
   - [#1138](https://github.com/wix-incubator/rich-content/pull/1138) plugin design overhaul
 - `text-selection-toolbar`